moved usfm-parse.jsfile to js folder and changed directory path in fetchData.js for usfm parser file
